Move StoryPanel to core, refactor editor/Visualizer
