--- web/src/beta/features/Editor/Visualizer/CanvasArea/index.tsx
@@ -1,119 +1,0 @@
-import React, { ReactNode, useCallback } from "react";
-
-import ContentPicker from "@reearth/beta/components/ContentPicker";
-import Visualizer, { type Props as VisualizerProps } from "@reearth/beta/lib/core/Visualizer";
-import { config } from "@reearth/services/config";
-
-import FovSlider from "../FovSlider";
-
-import useHooks from "./hooks";
-
-export type Props = {
-  sceneId?: string;
-  isBuilt?: boolean;
-  inEditor?: boolean;
-  children?: ReactNode;
-};
-
-const CanvasArea: React.FC<Props> = ({ sceneId, isBuilt, inEditor, children }) => {
-  const {
-    rootLayerId,
-    selectedBlockId,
-    zoomedLayerId,
-    sceneProperty,
-    pluginProperty,
-    clusters,
-    layers,
-    widgets,
-    tags,
-    selectedLayerId,
-    blocks,
-    isCapturing,
-    sceneMode,
-    camera,
-    selectedWidgetArea,
-    widgetAlignEditorActivated,
-    engineMeta,
-    layerSelectionReason,
-    useExperimentalSandbox,
-    isVisualizerReady: _isVisualizerReady,
-    selectLayer,
-    selectBlock,
-    onBlockChange,
-    onBlockMove,
-    onBlockRemove,
-    onBlockInsert,
-    onWidgetUpdate,
-    onWidgetAlignSystemUpdate,
-    selectWidgetArea,
-    onIsCapturingChange,
-    onCameraChange,
-    onFovChange,
-    handleDropLayer,
-    zoomToLayer,
-    handleMount,
-  } = useHooks({ sceneId, isBuilt });
-  const renderInfoboxInsertionPopUp = useCallback<
-    NonNullable<VisualizerProps["renderInfoboxInsertionPopup"]>
-  >(
-    (onSelect, onClose) => (
-      <ContentPicker items={blocks} onSelect={onSelect} onClickAway={onClose} />
-    ),
-    [blocks],
-  );
-
-  return (
-    <>
-      <Visualizer
-        engine="cesium"
-        isEditable={!isBuilt}
-        isBuilt={!!isBuilt}
-        inEditor={!!inEditor}
-        layers={layers}
-        widgetAlignSystem={widgets?.alignSystem}
-        floatingWidgets={widgets?.floating}
-        widgetLayoutConstraint={widgets?.layoutConstraint}
-        ownBuiltinWidgets={widgets?.ownBuiltinWidgets}
-        selectedLayerId={selectedLayerId}
-        selectedBlockId={selectedBlockId}
-        selectedWidgetArea={selectedWidgetArea}
-        zoomedLayerId={zoomedLayerId}
-        rootLayerId={rootLayerId}
-        sceneProperty={sceneProperty}
-        tags={tags}
-        pluginProperty={pluginProperty}
-        clusters={clusters}
-        camera={camera}
-        ready={isBuilt || (!!layers && !!widgets)}
-        pluginBaseUrl={config()?.plugins}
-        widgetAlignSystemEditing={widgetAlignEditorActivated}
-        meta={engineMeta}
-        layerSelectionReason={layerSelectionReason}
-        useExperimentalSandbox={useExperimentalSandbox}
-        onLayerSelect={selectLayer}
-        onCameraChange={onCameraChange}
-        onWidgetLayoutUpdate={onWidgetUpdate}
-        onWidgetAlignmentUpdate={onWidgetAlignSystemUpdate}
-        onWidgetAreaSelect={selectWidgetArea}
-        onBlockSelect={selectBlock}
-        onBlockChange={onBlockChange}
-        onBlockMove={onBlockMove}
-        onBlockDelete={onBlockRemove}
-        onBlockInsert={onBlockInsert}
-        onLayerDrop={handleDropLayer}
-        onZoomToLayer={zoomToLayer}
-        onMount={handleMount}
-        renderInfoboxInsertionPopup={renderInfoboxInsertionPopUp}>
-        {children}
-      </Visualizer>
-      <FovSlider
-        visible={isCapturing && sceneMode && sceneMode !== "2d"}
-        onIsCapturingChange={onIsCapturingChange}
-        camera={camera}
-        onFovChange={onFovChange}
-      />
-    </>
-  );
-};
-
-export default CanvasArea;
