+++ web/src/beta/features/Editor/Visualizer/hooks.ts
@@ -208,6 +208,7 @@ export default ({ sceneId, isBuilt }: { sceneId?: string; isBuilt?: boolean }) =
     layerSelectionReason,
     useExperimentalSandbox,
     isVisualizerReady,
+    selectWidgetArea: selectedWidgetAreaVar,
     selectLayer,
     selectBlock,
     onBlockChange,
@@ -215,7 +216,6 @@ export default ({ sceneId, isBuilt }: { sceneId?: string; isBuilt?: boolean }) =
     onBlockRemove,
     onBlockInsert,
     onWidgetUpdate,
-    selectWidgetArea: selectedWidgetAreaVar,
     onWidgetAlignSystemUpdate,
     onIsCapturingChange,
     onCameraChange,
