+++ web/src/beta/features/Editor/index.tsx
@@ -1,7 +1,6 @@
 import React, { useState } from "react";
 
 import Resizable from "@reearth/beta/components/Resizable";
-import StoryPanel from "@reearth/beta/features/Editor/StoryPanel";
 import useLeftPanel from "@reearth/beta/features/Editor/useLeftPanel";
 import useRightPanel from "@reearth/beta/features/Editor/useRightPanel";
 import useSecondaryNavbar from "@reearth/beta/features/Editor/useSecondaryNavbar";
@@ -129,19 +128,16 @@ const Editor: React.FC<Props> = ({ sceneId, projectId, workspaceId, tab, stories
               tab={tab}
               hasNav={!!secondaryNavbar}
               visualizerWidth={visualizerWidth}>
-              <Visualizer sceneId={sceneId}>
-                {selectedProjectType === "story" && (
-                  <StoryPanel
-                    sceneId={sceneId}
-                    selectedStory={selectedStory}
-                    currentPage={currentPage}
-                    isAutoScrolling={isAutoScrolling}
-                    installableBlocks={installableStoryBlocks}
-                    onAutoScrollingChange={handleAutoScrollingChange}
-                    onCurrentPageChange={handleCurrentPageChange}
-                  />
-                )}
-              </Visualizer>
+              <Visualizer
+                sceneId={sceneId}
+                showStoryPanel={selectedProjectType === "story"}
+                selectedStory={selectedStory}
+                currentPage={currentPage}
+                isAutoScrolling={isAutoScrolling}
+                installableBlocks={installableStoryBlocks}
+                onAutoScrollingChange={handleAutoScrollingChange}
+                onCurrentPageChange={handleCurrentPageChange}
+              />
             </VisualizerWrapper>
           </Center>
           {rightPanel && (
