+++ web/src/beta/lib/core/StoryPanel/Block/Template.tsx
