+++ web/src/beta/lib/core/StoryPanel/Block/builtin/Image/index.tsx
