+++ web/src/beta/lib/core/StoryPanel/PanelContent/index.tsx
